added 3 state button
added a 3 state button to be used on the override screen (example in debugging screen), Started to arrange objects in the override screen. The relays will be listed in a 2x8 grid of 3 state buttons (on, off, program controlled). The override states will be a 3x3 toggle grid (only 1 "on" at a time)